Move loaders to own packages
