+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/ExtensionDefinitionGenerator.java
@@ -37,9 +37,9 @@ import org.hl7.fhir.convertors.advisors.impl.BaseAdvisor_30_40;
 import org.hl7.fhir.convertors.factory.VersionConvertorFactory_10_40;
 import org.hl7.fhir.convertors.factory.VersionConvertorFactory_14_40;
 import org.hl7.fhir.convertors.factory.VersionConvertorFactory_30_40;
-import org.hl7.fhir.convertors.loaders.R2016MayToR4Loader;
-import org.hl7.fhir.convertors.loaders.R2ToR4Loader;
-import org.hl7.fhir.convertors.loaders.R3ToR4Loader;
+import org.hl7.fhir.convertors.loaders.loaderR4.R2016MayToR4Loader;
+import org.hl7.fhir.convertors.loaders.loaderR4.R2ToR4Loader;
+import org.hl7.fhir.convertors.loaders.loaderR4.R3ToR4Loader;
 import org.hl7.fhir.convertors.misc.IGR2ConvertorAdvisor;
 import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.r4.conformance.ProfileUtilities;
