+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/R2R3ConversionManager.java
@@ -29,7 +29,7 @@ package org.hl7.fhir.convertors;
   
  */
 
-import org.hl7.fhir.convertors.loaders.R2ToR3Loader;
+import org.hl7.fhir.convertors.loaders.loaderR3.R2ToR3Loader;
 import org.hl7.fhir.dstu3.context.SimpleWorkerContext;
 import org.hl7.fhir.dstu3.elementmodel.Manager;
 import org.hl7.fhir.dstu3.elementmodel.Manager.FhirFormat;
