+++ org.hl7.fhir.validation/src/test/java/org/hl7/fhir/conversion/tests/R3R4ConversionTests.java
@@ -17,7 +17,7 @@ import java.util.zip.ZipInputStream;
 
 import javax.xml.parsers.ParserConfigurationException;
 
-import org.hl7.fhir.convertors.loaders.R3ToR4Loader;
+import org.hl7.fhir.convertors.loaders.loaderR4.R3ToR4Loader;
 import org.hl7.fhir.exceptions.DefinitionException;
 import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.exceptions.FHIRFormatError;
