+++ pom.xml
@@ -19,7 +19,7 @@
 
     <properties>
         <hapi_fhir_version>5.1.0</hapi_fhir_version>
-        <validator_test_case_version>1.1.70</validator_test_case_version>
+        <validator_test_case_version>1.1.71-SNAPSHOT</validator_test_case_version>
         <junit_jupiter_version>5.7.1</junit_jupiter_version>
         <junit_platform_launcher_version>1.7.1</junit_platform_launcher_version>
         <maven_surefire_version>3.0.0-M4</maven_surefire_version>
